refactor(ui): разделён controller source management screen
SourceManagementScreenController оставлен координатором screenState и presented scenario flow
launch routing, add-feed, create-folder, move-source и status mapping вынесены в focused файлы
